% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AUCO}
\alias{AUCO}
\title{Illapel Ecological Dataset}
\format{
A data frame with rodent and rainfall data.
}
\source{
Las Chinchillas National Reserve Research Station
}
\usage{
AUCO
}
\description{
Raw rodent and rainfall data collected from the Las Chinchillas
National Reserve near Illapel, Coquimbo Region of Chile. This dataset provides
ecological time series for studying species interactions and environmental effects.
}
\details{
Illapel Ecological Dataset
}
\examples{
data(AUCO)
head(AUCO)
summary(AUCO)
}
\keyword{datasets}
