% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_indices}
\alias{climate_indices}
\title{Climate index}
\format{
An object of class \code{data.frame} with 535 rows and 5 columns.
}
\usage{
climate_indices
}
\description{
This is a time series of climate index.
}
\examples{
head(climate_indices)
}
\keyword{datasets}
