% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureOfCausality.R
\name{natureOfCausality}
\alias{natureOfCausality}
\title{Causality Nature Classifier}
\usage{
natureOfCausality(PC, dur, hashedpatterns, len)
}
\arguments{
\item{PC}{Array, a three-dimensional pattern causality matrix where each cell represents the causal influence
from one variable to another over time.}

\item{dur}{Integer vector, the indices of time steps at which to evaluate causality.}

\item{hashedpatterns}{Numeric vector, typically an indexed representation of the variables or patterns under analysis.}

\item{len}{Integer, the expected length of the output vectors for each type of causality.}
}
\value{
A dataframe with four columns: \code{noCausality}, \code{Positive}, \code{Negative}, and \code{Dark}, each indicating
the presence (1) or absence (0) of the respective type of causality at each time step.
}
\description{
Processes a 3-dimensional pattern causality matrix to identify and classify the nature of causal
relationships over a specified duration. It determines if the causality is positive, negative, dark, or nonexistent
based on the matrix values.
}
\details{
Classify the Nature of Causality from a Pattern Causality Matrix

This function analyzes the pattern causality matrix over a given duration to classify the nature of causality
(positive, negative, dark, or no causality) between variables in a dynamic complex system. The function checks
for causal relationships based on the entries in the pattern causality matrix (PC) and categorizes the nature
of these interactions.
}
\examples{
# Example usage:
PC <- array(runif(100), dim = c(10, 10, 10))
dur <- 1:10
hashedpatterns <- seq(1, 10)
len <- 10
result <- natureOfCausality(PC, dur, hashedpatterns, len)
}
