% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastNNsInfo.R
\name{pastNNsInfo}
\alias{pastNNsInfo}
\title{Retrieve Information on Past Nearest Neighbors}
\usage{
pastNNsInfo(CCSPAN, NNSPAN, Mx, Dx, SMx, PSMx, i, h)
}
\arguments{
\item{CCSPAN}{Integer, the span of common coordinates to exclude in the nearest neighbor search.}

\item{NNSPAN}{Integer, the number of nearest neighbors to consider for the analysis.}

\item{Mx}{Matrix, the main matrix representing the state space of the system.}

\item{Dx}{Numeric matrix, containing distances between points in the state space.}

\item{SMx}{Matrix, containing signatures of the state space.}

\item{PSMx}{Matrix, containing patterns derived from the signatures.}

\item{i}{Integer, the current index in time series data for which nearest neighbors are being considered.}

\item{h}{Integer, the horizon beyond which data is not considered in the nearest neighbor search.}
}
\value{
A list containing indices, times, distances, signatures, patterns, and original coordinates of the nearest
neighbors from past data, useful for subsequent analysis and prediction in the context of complex systems.
}
\description{
This function extracts and organizes information about the nearest neighbors of a given point in time series data,
utilizing past data up to a specified horizon. It operates within a framework that incorporates multiple matrices
representing different attributes of the system's state space, such as distances, signatures, and patterns.
}
\examples{
# Random data generation for demonstration
set.seed(123)
Mx <- matrix(rnorm(200), nrow=20)
Dx <- as.matrix(dist(Mx))
SMx <- apply(Mx, 1, function(x) diff(x))
PSMx <- apply(SMx, 1, function(x) ifelse(x > 0, 1, ifelse(x < 0, -1, 0)))
CCSPAN <- 5
NNSPAN <- 3
i <- 15
h <- 2
neighborsInfo <- pastNNsInfo(CCSPAN, NNSPAN, Mx, Dx, SMx, PSMx, i, h)
print(neighborsInfo)
}
