% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternHashing.R
\name{patternHashing}
\alias{patternHashing}
\title{Pattern Hashing function}
\usage{
patternHashing(E)
}
\arguments{
\item{E}{the embedded dimensions}
}
\description{
Appointing Unique Identifiers to Symbolic Patterns
}
