% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{pc_cv}
\alias{pc_cv}
\title{Pattern Causality Cross-Validation Object}
\usage{
pc_cv(samples = NULL, results = NULL, summary = NULL, parameters = NULL)
}
\arguments{
\item{samples}{Numeric vector of sample sizes used.}

\item{results}{Matrix containing causality results for each sample.}

\item{summary}{Matrix of summary statistics.}

\item{parameters}{List of analysis parameters.}
}
\value{
An object of class "pc_cv".
}
\description{
Creates a pattern causality cross-validation object containing results
from repeated sampling analysis.
This function constructs an object of class \code{pc_cv} to store the results
of cross-validation analysis.
}
