% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{pc_effect}
\alias{pc_effect}
\title{Pattern Causality Effect Object}
\usage{
pc_effect(positive = NULL, negative = NULL, dark = NULL, items = NULL)
}
\arguments{
\item{positive}{Data frame containing positive causality effects.}

\item{negative}{Data frame containing negative causality effects.}

\item{dark}{Data frame containing dark causality effects.}

\item{items}{Names of items in the analysis.}
}
\value{
An object of class "pc_effect".
}
\description{
Creates a pattern causality effect object that contains information about
the received and exerted influences for different causality types.
This function constructs an object of class \code{pc_effect} to store the results
of effect analysis.
}
