% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{pc_fit}
\alias{pc_fit}
\title{Pattern Causality Fit Object}
\usage{
pc_fit(total = NULL, positive = NULL, negative = NULL, dark = NULL)
}
\arguments{
\item{total}{Total causality strength}

\item{positive}{Positive causality strength}

\item{negative}{Negative causality strength}

\item{dark}{Dark causality strength}
}
\value{
An object of class "pc_fit"
}
\description{
Create a pattern causality fit object
}
