% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_params.R
\name{pc_params}
\alias{pc_params}
\title{Pattern Causality Parameter Optimization Results}
\usage{
pc_params(accuracy_summary, computation_time, parameters)
}
\arguments{
\item{accuracy_summary}{Data frame containing accuracy results for different parameter combinations}

\item{computation_time}{Time taken for optimization}

\item{parameters}{List of optimization parameters}
}
\value{
An object of class "pc_params"
}
\description{
Creates an object containing parameter optimization results for pattern causality analysis
}
\details{
Pattern Causality Parameter Optimization Results
}
