% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot.pc_fit}
\alias{plot.pc_fit}
\title{Plot Pattern Causality Results}
\usage{
\method{plot}{pc_fit}(x, ...)
}
\arguments{
\item{x}{A \code{pc_fit} object.}

\item{...}{Additional arguments passed to the underlying plotting functions.}
}
\value{
NULL invisibly.
}
\description{
Generates a combined plot of total causality and causality components for a \code{pc_fit} object.
This function combines the visualizations from \code{plot_total} and \code{plot_components} into a single plot.
}
