% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot_components.pc_fit}
\alias{plot_components.pc_fit}
\title{Plot Causality Components}
\usage{
\method{plot_components}{pc_fit}(x, ...)
}
\arguments{
\item{x}{A pc_fit object}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A ggplot object
}
\description{
Plot Causality Components
}
