% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot_components}
\alias{plot_components}
\title{Plot Causality Components}
\usage{
plot_components(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
Plot Causality Components
}
