% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot_total.pc_fit}
\alias{plot_total.pc_fit}
\title{Plot Total Causality}
\usage{
\method{plot_total}{pc_fit}(x, ...)
}
\arguments{
\item{x}{A pc_fit object}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A ggplot object
}
\description{
Plot Total Causality
}
