% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.pc_cv}
\alias{print.pc_cv}
\title{Print Pattern Causality Cross Validation Results}
\usage{
\method{print}{pc_cv}(x, ...)
}
\arguments{
\item{x}{A \code{pc_cv} object.}

\item{...}{Additional arguments passed to the \code{print} function.}
}
\value{
Invisibly returns the input object.
}
\description{
Prints the pattern causality cross-validation results.
This function displays the parameters used for cross-validation,
the sample sizes, and the summary statistics.
}
\examples{
data(climate_indices)
X <- climate_indices$AO
Y <- climate_indices$AAO
numberset <- c(100, 150, 200)
cv_results <- pcCrossValidation(X, Y, 3, 2, "euclidean", 1, FALSE, numberset = numberset)
print(cv_results)
}
