% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.pc_fit}
\alias{print.pc_fit}
\title{Print Pattern Causality Results}
\usage{
\method{print}{pc_fit}(x, ...)
}
\arguments{
\item{x}{A \code{pc_fit} object.}

\item{...}{Additional arguments passed to the \code{print} function.}
}
\value{
Invisibly returns the input object.
}
\description{
Prints the pattern causality analysis results from a \code{pc_fit} object.
This function displays the total, positive, negative, and dark causality strengths.
}
