% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_params.R
\name{print.pc_params}
\alias{print.pc_params}
\title{Print Method for Pattern Causality Parameter Results}
\usage{
\method{print}{pc_params}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A pc_params object}

\item{verbose}{Logical; whether to display detailed information}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for Pattern Causality Parameter Results
}
