% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.pc_state}
\alias{print.pc_state}
\title{Print State Space Reconstruction}
\usage{
\method{print}{pc_state}(x, ...)
}
\arguments{
\item{x}{A \code{pc_state} object.}

\item{...}{Additional arguments passed to the \code{print} function.}
}
\value{
Invisibly returns the input object.
}
\description{
Prints the state space reconstruction results.
This function displays the parameters used for state space reconstruction
and a preview of the reconstructed points.
}
