% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock}
\alias{stock}
\title{Stock index}
\format{
An object of class \code{data.frame} with 8167 rows and 12 columns.
}
\usage{
stock
}
\description{
This is a time series of stock series which contains AAPL and MSFT from March 13, 1986 to August 6, 2018.
}
\examples{
head(stock)
}
\keyword{datasets}
