% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_accuracy_methods.R
\name{summary.pc_accuracy}
\alias{summary.pc_accuracy}
\title{Summary Method for Pattern Causality Accuracy Results}
\usage{
\method{summary}{pc_accuracy}(object, ...)
}
\arguments{
\item{object}{A pc_accuracy object}

\item{...}{Additional arguments passed to summary}
}
\value{
A summary object for pc_accuracy
}
\description{
Summary Method for Pattern Causality Accuracy Results
}
