% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.pc_fit}
\alias{summary.pc_fit}
\title{Summarize Pattern Causality Results}
\usage{
\method{summary}{pc_fit}(object, ...)
}
\arguments{
\item{object}{A pc_fit object}

\item{...}{Additional arguments passed to summary}
}
\value{
Invisibly returns the input object
}
\description{
Summarize Pattern Causality Results
}
