% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_params.R
\name{summary.pc_params}
\alias{summary.pc_params}
\title{Summary Method for Pattern Causality Parameter Results}
\usage{
\method{summary}{pc_params}(object, ...)
}
\arguments{
\item{object}{A pc_params object}

\item{...}{Additional arguments passed to summary}
}
\value{
A summary object for pc_params
}
\description{
Summary Method for Pattern Causality Parameter Results
}
