% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AUCO}
\alias{AUCO}
\title{A data from Illapel}
\format{
An object of class \code{data.frame} with 32 rows and 5 columns.
}
\usage{
AUCO
}
\description{
Raw rodent and rainfall data from the study site at Las Chinchillas National Reserve near the city of Illapel, Coquimbo Region of Chil.
}
\examples{
head(AUCO)
}
\keyword{datasets}
