% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DJS}
\alias{DJS}
\title{Dow Jones Stock Price}
\format{
An object of class \code{data.frame} with 4510 rows and 30 columns.
}
\usage{
DJS
}
\description{
It includes 29 time series of stock series in Dow Jones.
}
\examples{
head(DJS)
}
\keyword{datasets}
