% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot.pc_fit}
\alias{plot.pc_fit}
\title{Plot Pattern Causality Results}
\usage{
\method{plot}{pc_fit}(x, ...)
}
\arguments{
\item{x}{A pc_fit object}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A list of two ggplot objects
}
\description{
Plot Pattern Causality Results
}
