% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot_components.pc_fit}
\alias{plot_components.pc_fit}
\title{Plot Causality Components}
\usage{
\method{plot_components}{pc_fit}(x, ...)
}
\arguments{
\item{x}{A \code{pc_fit} object.}

\item{...}{Additional arguments passed to the underlying plotting functions.}
}
\value{
NULL.
}
\description{
Visualizes the positive, negative, and dark causality components as a barplot for a \code{pc_fit} object.
This function generates a barplot showing the strength of each causality component.
}
