% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot_total.pc_fit}
\alias{plot_total.pc_fit}
\title{Plot Total Causality}
\usage{
\method{plot_total}{pc_fit}(x, ...)
}
\arguments{
\item{x}{A \code{pc_fit} object.}

\item{...}{Additional arguments passed to the underlying plotting functions.}
}
\value{
NULL.
}
\description{
Visualizes the total causality strength as a barplot for a \code{pc_fit} object.
This function generates a barplot showing the total causality strength and its complement.
}
