% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.pc_fit}
\alias{print.pc_fit}
\title{Print Pattern Causality Results}
\usage{
\method{print}{pc_fit}(x, ...)
}
\arguments{
\item{x}{A pc_fit object}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print Pattern Causality Results
}
