% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.pc_fit}
\alias{summary.pc_fit}
\title{Summarize Pattern Causality Results}
\usage{
\method{summary}{pc_fit}(object, ...)
}
\arguments{
\item{object}{A \code{pc_fit} object.}

\item{...}{Additional arguments passed to the \code{summary} function.}
}
\value{
Invisibly returns the input object.
}
\description{
Provides a summary of the pattern causality analysis results from a \code{pc_fit} object.
This function displays a table of causality strengths for total, positive, negative, and dark components.
}
